% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_OTU.R
\name{clean_OTU}
\alias{clean_OTU}
\title{Clean the abundance data}
\usage{
clean_OTU(
  abundance_table = "Galaxy21-[FROGS_BIOM_to_TSV__abundance.tsv].tsv",
  affiliation_table = "Galaxy22-[FROGS_BIOM_to_TSV__multi-affiliations.tsv].tsv",
  thresh = 0.5,
  name = "2022JRLAT-16S"
)
}
\arguments{
\item{abundance_table}{abundance table}

\item{affiliation_table}{affiliation table}

\item{thresh}{\% threshold below which 'unknown' are ignored}

\item{name}{name will be used to import files and write names before extension of the different output tables}
}
\description{
Work on raw abundance and multihits text files from Frogs.

Correct the taxonomy of each cluster of the abundance tsv file tagged 'multi-affiliation', from decision rules applied on the cleaned multihits tsv file:

Clean the taxonomy (7 ranks: Kingdom,Phylum,Class,Order,Family,Genus,Species) of each hit of the two files for:
\itemize{
\item the special characters '\url{}_"0123456789.' in all 7 taxonomic ranks.
\item 'mitochondria' (in the Family rank) and 'chloroplast' (in the Class rank), which are replaced by plantae in all 7 taxonomic ranks.
\item 'unidentified', 'sp.' and 'bacterium', which are replaced by 'unknown species' in the Species rank .
\item unexpected number of words in all 7 taxonomic ranks.
\item for the species and genus ranks only: if the number of 'unknown species' hits < \% threshold then ignore them.
\item for all taxonomic ranks:
\item if there is a unique taxonomic name, then rename by this taxonomic name.
\item if there is a unique taxonomic name and an 'unknown', then rename by 'unknown'.
\item if there are two or more taxonomic names, then name 'multi-affiliation'.
Identify chimeric sequences using the isBimeraDenovo(...) function in the dada2 library
Export an abundance table tsv file cleaned for taxonomy and filtered for chimeras.
}
}
