% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_chimeras.R
\name{find_chimeras}
\alias{find_chimeras}
\title{Find chimeras}
\usage{
find_chimeras(fordada, name)
}
\arguments{
\item{fordada}{abundance table}
}
\description{
Functions that finds and cleans chimeras
}
