% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_Binomial.R
\name{sample_Binomial}
\alias{sample_Binomial}
\title{Sampling function for the abundances of species in different habitats}
\usage{
sample_Binomial(Nhabitats, prob, mean_nindiviuals_per_habitat)
}
\arguments{
\item{Nhabitats}{number of habitats (could be either a vector of values or a single value that will be recycle)}

\item{prob}{frequency of OTU1 (could be either a vector of values or a single value that will be recycle)}

\item{mean_nindiviuals_per_habitat}{average number of individuals per habitat (could be either a vector of values or a single value that will be recycle)}
}
\value{
dataframe
}
\description{
Functions that randomly sample abundance data in N habitat and return the abundance table
}
