% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_Multinomial.R
\name{sample_Multinomial}
\alias{sample_Multinomial}
\title{Sampling function for the abundances of species in different habitats}
\usage{
sample_Multinomial(Nsamples, prob, mean_nindiviuals_per_habitat)
}
\arguments{
\item{Nsamples}{number of samples (could be either a vector of values or a single value that will be recycle)}

\item{prob}{frequency of all OTUs (could be either a vector of values or a single value that will be recycle)}

\item{mean_nindiviuals_per_habitat}{average number of individuals per habitat (could be either a vector of values or a single value that will be recycle)}

\item{seed}{seed used to perform simulations}
}
\value{
dataframe
}
\description{
Functions that randomly sample abundance data in N habitat and return the abundance table
}
