% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_Poisson.R
\name{sample_Poisson}
\alias{sample_Poisson}
\title{Sampling function for Poisson data}
\usage{
sample_Poisson(sample, count, lambda)
}
\arguments{
\item{sample}{count lambda}
}
\value{
dataframe
}
\description{
Functions that randomly samples Poisson data and returns a dataframe
}
